收益计划计算工具微调

+++ autotest-core/src/main/java/com/autotest/core/model/BfProductDetail.java
@@ -2,6 +2,7 @@ package com.autotest.core.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Data
@@ -9,9 +10,9 @@ public class BfProductDetail {
     private int id;
     private String productCode;
     private String productName;
-    private double interestRate;
-    private double monthRate;
-    private double falseInterestRate;
+    private BigDecimal interestRate;
+    private BigDecimal monthRate;
+    private BigDecimal falseInterestRate;
     private short term;
     private String refundWay;
     private String channel;
@@ -22,9 +23,9 @@ public class BfProductDetail {
     private String productBusinessName;
     private short advancedDays;
     private short advancedRules;
-    private double delayRate;
+    private BigDecimal delayRate;
     private int delay_no_interest_days;
-    private double advance_cost;
-    private double advance_cost60;
-    private double manage_cost;
+    private BigDecimal advance_cost;
+    private BigDecimal advance_cost60;
+    private BigDecimal manage_cost;
 }
